fix: address code review findings
- Change setup_tcp_pair cdc_itf parameter from uint8_t to int so the -1
  sentinel for optional connections propagates cleanly without a cast at the
  call site; document the uint8_t storage in the comment inside the function
- Remove the now-unnecessary (uint8_t) cast in connect_and_setup_pair

Agent-Logs-Url: https://github.com/blue-wind-25/JxMake/sessions/ec1fd2b1-4276-432c-aeed-44e1b2216a4b

+++ hardware/Tools/Firmware/Serial_WiFi_Bridge/PicoW/src/main.c
@@ -266,7 +266,7 @@ static struct tcp_pcb* connect_and_setup_pair(
         cleanupAndBlinkErrorLED(err_led);
     }
 
-    setup_tcp_pair(pair, pcb, (uint8_t) cdc_itf, tx_storage, rx_storage, buf_size, tx_lock, rx_lock);
+    setup_tcp_pair(pair, pcb, cdc_itf, tx_storage, rx_storage, buf_size, tx_lock, rx_lock);
     if(error_cb) tcp_err(pcb, error_cb);
 
     const uint32_t svcTimeoutMS = SERVER_CONNECT_TIMEOUT_MS;
